+++ app/src/main/java/com/depromeet/housekeeper/di/NetworkModule.kt
@@ -2,15 +2,20 @@ package com.depromeet.housekeeper.di
 
 import com.depromeet.housekeeper.data.ApiService
 import com.depromeet.housekeeper.data.dataSource.RemoteDataSourceImpl
+import com.depromeet.housekeeper.util.NETWORK_ERROR
+import com.depromeet.housekeeper.util.PrefsManager
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.Dispatchers
-import okhttp3.OkHttpClient
+import okhttp3.*
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.converter.moshi.MoshiConverterFactory
 import java.util.concurrent.TimeUnit
 import javax.inject.Singleton
 
@@ -19,27 +24,52 @@ import javax.inject.Singleton
 class NetworkModule {
     private val BASE_URL = "http://ec2-3-39-60-64.ap-northeast-2.compute.amazonaws.com:8080"
 
-    @Singleton
-    @Provides
-    fun provideOkHttpBuilder(): OkHttpClient {
-        val httpLoggingInterceptor = HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
+    private val networkInterceptor: Interceptor = Interceptor { chain ->
+        val request = chain.request()
+        try {
+            chain.proceed(
+                request.newBuilder()
+                    .addHeader(
+                        "Authorization",
+                        PrefsManager.refreshToken.ifEmpty { PrefsManager.authCode })
+                    .build()
+            )
+        } catch (e: Exception) {
+            Response.Builder()
+                .request(request)
+                .protocol(Protocol.HTTP_1_1)
+                .code(NETWORK_ERROR)
+                .message(e.message ?: "")
+                .body(ResponseBody.create(null, e.message ?: ""))
+                .build()
         }
-        return OkHttpClient.Builder()
-            .addNetworkInterceptor(httpLoggingInterceptor)
-            .connectTimeout(5, TimeUnit.SECONDS)
-            .readTimeout(5, TimeUnit.SECONDS)
-            .writeTimeout(5, TimeUnit.SECONDS)
-            .build()
     }
 
     @Singleton
     @Provides
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
+    fun provideOkHttpBuilder(): OkHttpClient.Builder {
+        val httpLoggingInterceptor = HttpLoggingInterceptor().apply {
+            level = HttpLoggingInterceptor.Level.BODY
+        }
+        return OkHttpClient.Builder()
+            .addInterceptor(networkInterceptor)
+            .addNetworkInterceptor(httpLoggingInterceptor)
+            .connectTimeout(5, TimeUnit.SECONDS)
+            .readTimeout(5, TimeUnit.SECONDS)
+            .writeTimeout(5, TimeUnit.SECONDS)
+    }
+
+    @Singleton
+    @Provides
+    fun provideRetrofit(okHttpClientBuilder: OkHttpClient.Builder): Retrofit {
+        val moshi = Moshi.Builder()
+            .add(KotlinJsonAdapterFactory())
+            .build()
         return Retrofit.Builder()
             .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(okHttpClient)
+            .client(okHttpClientBuilder.build())
+            .addConverterFactory(MoshiConverterFactory.create(moshi).asLenient())
+            .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .build()
     }
 
@@ -54,5 +84,4 @@ class NetworkModule {
     fun provideRemoteDataSource(apiService: ApiService): RemoteDataSourceImpl {
         return RemoteDataSourceImpl(apiService, Dispatchers.IO)
     }
-
 }
