+++ app/src/main/java/com/depromeet/housekeeper/ui/join/InviteViewModel.kt
@@ -2,7 +2,6 @@ package com.depromeet.housekeeper.ui.join
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.depromeet.housekeeper.data.repository.Repository
 import com.depromeet.housekeeper.data.repository.UserRepository
 import com.depromeet.housekeeper.model.request.BuildTeam
 import com.depromeet.housekeeper.model.enums.InviteViewType
